Reorganize Windows-specific outputs
So that the various Windows imports aren't required. Also
don't allow the --pipe option on non-Windows platforms.
